some theme refactor
