+++ package.json
@@ -10,7 +10,7 @@
     "vercel:install": "yarn set version berry && yarn install",
     "native": "cd apps/expo && yarn start",
     "watch": "yarn workspaces foreach -pi run watch",
-    "web": "yarn build && cd apps/next && yarn next",
+    "web": "yarn build && cd apps/next && yarn dev",
     "web:extract": "DISABLE_EXTRACTION=false yarn workspace next-app dev",
     "web:prod": "yarn workspace next-app build",
     "web:prod:serve": "yarn workspace next-app serve",
