--- packages/ui/src/themes/colors-neon.ts
@@ -1,264 +1,0 @@
-export const blueDark = {
-  blue1: 'hsl(212, 100%, 9.2%)',
-  blue2: 'hsl(216, 100%, 11.8%)',
-  blue3: 'hsl(214, 100%, 15.3%)',
-  blue4: 'hsl(214, 100%, 17.9%)',
-  blue5: 'hsl(213, 100%, 20.2%)',
-  blue6: 'hsl(212, 100%, 23.1%)',
-  blue7: 'hsl(211, 100%, 27.4%)',
-  blue8: 'hsl(211, 100%, 34.1%)',
-  blue9: 'hsl(206, 100%, 50.0%)',
-  blue10: 'hsl(209, 100%, 60.6%)',
-  blue11: 'hsl(210, 100%, 66.1%)',
-  blue12: 'hsl(206, 100%, 85.8%)',
-}
-export const grayDark = {
-  gray1: 'hsl(0, 0%, 8.5%)',
-  gray2: 'hsl(0, 0%, 11.0%)',
-  gray3: 'hsl(0, 0%, 13.6%)',
-  gray4: 'hsl(0, 0%, 15.8%)',
-  gray5: 'hsl(0, 0%, 17.9%)',
-  gray6: 'hsl(0, 0%, 20.5%)',
-  gray7: 'hsl(0, 0%, 24.3%)',
-  gray8: 'hsl(0, 0%, 31.2%)',
-  gray9: 'hsl(0, 0%, 43.9%)',
-  gray10: 'hsl(0, 0%, 49.4%)',
-  gray11: 'hsl(0, 0%, 62.8%)',
-  gray12: 'hsl(0, 0%, 93.0%)',
-}
-export const greenDark = {
-  green1: 'hsl(146, 100%, 7.4%)',
-  green2: 'hsl(155, 100%, 8.4%)',
-  green3: 'hsl(155, 100%, 10.9%)',
-  green4: 'hsl(154, 100%, 12.9%)',
-  green5: 'hsl(154, 100%, 14.9%)',
-  green6: 'hsl(154, 100%, 17.6%)',
-  green7: 'hsl(153, 100%, 21.8%)',
-  green8: 'hsl(151, 100%, 28.4%)',
-  green9: 'hsl(151, 100%, 41.5%)',
-  green10: 'hsl(151, 100%, 46.5%)',
-  green11: 'hsl(151, 100%, 53.2%)',
-  green12: 'hsl(137, 100%, 84%)',
-}
-export const orangeDark = {
-  orange1: 'hsl(30, 100%, 7.2%)',
-  orange2: 'hsl(28, 100%, 8.4%)',
-  orange3: 'hsl(26, 100%, 11.6%)',
-  orange4: 'hsl(25, 100%, 14.1%)',
-  orange5: 'hsl(24, 100%, 16.6%)',
-  orange6: 'hsl(24, 100%, 19.8%)',
-  orange7: 'hsl(24, 100%, 24.0%)',
-  orange8: 'hsl(25, 100%, 29.0%)',
-  orange9: 'hsl(24, 100%, 50.0%)',
-  orange10: 'hsl(24, 100%, 58.5%)',
-  orange11: 'hsl(24, 100%, 62.2%)',
-  orange12: 'hsl(24, 100%, 83.2%)',
-}
-export const pinkDark = {
-  pink1: 'hsl(318, 100%, 9.6%)',
-  pink2: 'hsl(319, 100%, 11.6%)',
-  pink3: 'hsl(319, 100%, 16.0%)',
-  pink4: 'hsl(320, 100%, 18.7%)',
-  pink5: 'hsl(320, 100%, 21.1%)',
-  pink6: 'hsl(321, 100%, 24.4%)',
-  pink7: 'hsl(321, 100%, 29.7%)',
-  pink8: 'hsl(322, 100%, 37.5%)',
-  pink9: 'hsl(322, 100%, 54.5%)',
-  pink10: 'hsl(323, 100%, 59.2%)',
-  pink11: 'hsl(325, 100%, 66.4%)',
-  pink12: 'hsl(322, 100%, 85.8%)',
-}
-export const purpleDark = {
-  purple1: 'hsl(284, 100%, 9.6%)',
-  purple2: 'hsl(283, 100%, 11.8%)',
-  purple3: 'hsl(281, 100%, 16.5%)',
-  purple4: 'hsl(280, 100%, 20.0%)',
-  purple5: 'hsl(279, 100%, 23.3%)',
-  purple6: 'hsl(277, 100%, 27.5%)',
-  purple7: 'hsl(275, 100%, 34.6%)',
-  purple8: 'hsl(272, 100%, 45.9%)',
-  purple9: 'hsl(272, 100%, 54.0%)',
-  purple10: 'hsl(273, 100%, 59.1%)',
-  purple11: 'hsl(275, 100%, 71.0%)',
-  purple12: 'hsl(279, 100%, 85.7%)',
-}
-export const redDark = {
-  red1: 'hsl(353, 100%, 9.8%)',
-  red2: 'hsl(357, 100%, 12.0%)',
-  red3: 'hsl(356, 100%, 16.4%)',
-  red4: 'hsl(356, 100%, 19.2%)',
-  red5: 'hsl(356, 100%, 21.9%)',
-  red6: 'hsl(356, 100%, 25.9%)',
-  red7: 'hsl(357, 100%, 31.8%)',
-  red8: 'hsl(358, 100%, 40.4%)',
-  red9: 'hsl(358, 100%, 59.0%)',
-  red10: 'hsl(358, 100%, 64.0%)',
-  red11: 'hsl(358, 100%, 69.5%)',
-  red12: 'hsl(351, 100%, 86%)',
-}
-export const yellowDark = {
-  yellow1: 'hsl(45, 100%, 5.5%)',
-  yellow2: 'hsl(46, 100%, 6.7%)',
-  yellow3: 'hsl(45, 100%, 8.7%)',
-  yellow4: 'hsl(45, 100%, 10.4%)',
-  yellow5: 'hsl(47, 100%, 12.1%)',
-  yellow6: 'hsl(49, 100%, 14.3%)',
-  yellow7: 'hsl(49, 100%, 18.4%)',
-  yellow8: 'hsl(50, 100%, 22.0%)',
-  yellow9: 'hsl(53, 100%, 50.0%)',
-  yellow10: 'hsl(54, 100%, 68.0%)',
-  yellow11: 'hsl(48, 100%, 47.0%)',
-  yellow12: 'hsl(53, 100%, 81%)',
-}
-
-export const blue = {
-  blue1: 'hsl(206, 100%, 86.8%)',
-  blue2: 'hsl(210, 100%, 66.1%)',
-  blue3: 'hsl(209, 100%, 60.6%)',
-  blue4: 'hsl(206, 100%, 50.0%)',
-  blue5: 'hsl(211, 100%, 34.1%)',
-  blue6: 'hsl(211, 100%, 27.4%)',
-  blue7: 'hsl(212, 100%, 23.1%)',
-  blue8: 'hsl(213, 100%, 20.2%)',
-  blue9: 'hsl(214, 100%, 17.9%)',
-  blue10: 'hsl(214, 100%, 15.3%)',
-  blue11: 'hsl(216, 100%, 11.8%)',
-  blue12: 'hsl(212, 100%, 9.2%)',
-}
-export const gray = {
-  gray1: 'hsl(0, 0%, 99.0%)',
-  gray2: 'hsl(0, 0%, 97.3%)',
-  gray3: 'hsl(0, 0%, 95.1%)',
-  gray4: 'hsl(0, 0%, 93.0%)',
-  gray5: 'hsl(0, 0%, 90.9%)',
-  gray6: 'hsl(0, 0%, 88.7%)',
-  gray7: 'hsl(0, 0%, 85.8%)',
-  gray8: 'hsl(0, 0%, 78.0%)',
-  gray9: 'hsl(0, 0%, 56.1%)',
-  gray10: 'hsl(0, 0%, 52.3%)',
-  gray11: 'hsl(0, 0%, 43.5%)',
-  gray12: 'hsl(0, 0%, 9.0%)',
-}
-export const green = {
-  green1: 'hsl(137, 100%, 73%)',
-  green2: 'hsl(151, 100%, 53.2%)',
-  green3: 'hsl(151, 100%, 46.5%)',
-  green4: 'hsl(151, 100%, 41.5%)',
-  green5: 'hsl(151, 100%, 28.4%)',
-  green6: 'hsl(153, 100%, 21.8%)',
-  green7: 'hsl(154, 100%, 17.6%)',
-  green8: 'hsl(154, 100%, 14.9%)',
-  green9: 'hsl(154, 100%, 12.9%)',
-  green10: 'hsl(155, 100%, 10.9%)',
-  green11: 'hsl(155, 100%, 8.4%)',
-  green12: 'hsl(146, 100%, 7.4%)',
-}
-export const orange = {
-  orange1: 'hsl(24, 100%, 82.2%)',
-  orange2: 'hsl(24, 100%, 62.2%)',
-  orange3: 'hsl(24, 100%, 58.5%)',
-  orange4: 'hsl(24, 100%, 50.0%)',
-  orange5: 'hsl(25, 100%, 29.0%)',
-  orange6: 'hsl(24, 100%, 24.0%)',
-  orange7: 'hsl(24, 100%, 19.8%)',
-  orange8: 'hsl(24, 100%, 16.6%)',
-  orange9: 'hsl(25, 100%, 14.1%)',
-  orange10: 'hsl(26, 100%, 11.6%)',
-  orange11: 'hsl(28, 100%, 8.4%)',
-  orange12: 'hsl(30, 100%, 7.2%)',
-}
-export const pink = {
-  pink1: 'hsl(322, 100%, 86.8%)',
-  pink2: 'hsl(325, 100%, 66.4%)',
-  pink3: 'hsl(323, 100%, 59.2%)',
-  pink4: 'hsl(322, 100%, 54.5%)',
-  pink5: 'hsl(322, 100%, 37.5%)',
-  pink6: 'hsl(321, 100%, 29.7%)',
-  pink7: 'hsl(321, 100%, 24.4%)',
-  pink8: 'hsl(320, 100%, 21.1%)',
-  pink9: 'hsl(320, 100%, 18.7%)',
-  pink10: 'hsl(319, 100%, 16.0%)',
-  pink11: 'hsl(319, 100%, 11.6%)',
-  pink12: 'hsl(318, 100%, 9.6%)',
-}
-export const purple = {
-  purple1: 'hsl(279, 100%, 91.7%)',
-  purple2: 'hsl(275, 100%, 71.0%)',
-  purple3: 'hsl(273, 100%, 59.1%)',
-  purple4: 'hsl(272, 100%, 54.0%)',
-  purple5: 'hsl(272, 100%, 45.9%)',
-  purple6: 'hsl(275, 100%, 34.6%)',
-  purple7: 'hsl(277, 100%, 27.5%)',
-  purple8: 'hsl(279, 100%, 23.3%)',
-  purple9: 'hsl(280, 100%, 20.0%)',
-  purple10: 'hsl(281, 100%, 16.5%)',
-  purple11: 'hsl(283, 100%, 11.8%)',
-  purple12: 'hsl(284, 100%, 9.6%)',
-}
-export const red = {
-  red1: 'hsl(351, 100%, 89%)',
-  red2: 'hsl(358, 100%, 69.5%)',
-  red3: 'hsl(358, 100%, 64.0%)',
-  red4: 'hsl(358, 100%, 59.0%)',
-  red5: 'hsl(358, 100%, 40.4%)',
-  red6: 'hsl(357, 100%, 31.8%)',
-  red7: 'hsl(356, 100%, 25.9%)',
-  red8: 'hsl(356, 100%, 21.9%)',
-  red9: 'hsl(356, 100%, 19.2%)',
-  red10: 'hsl(356, 100%, 16.4%)',
-  red11: 'hsl(357, 100%, 12.0%)',
-  red12: 'hsl(353, 100%, 9.8%)',
-}
-export const yellow = {
-  yellow1: 'hsl(53, 100%, 67%)',
-  yellow2: 'hsl(48, 100%, 47.0%)',
-  yellow3: 'hsl(54, 100%, 68.0%)',
-  yellow4: 'hsl(53, 100%, 50.0%)',
-  yellow5: 'hsl(50, 100%, 22.0%)',
-  yellow6: 'hsl(49, 100%, 18.4%)',
-  yellow7: 'hsl(49, 100%, 14.3%)',
-  yellow8: 'hsl(47, 100%, 12.1%)',
-  yellow9: 'hsl(45, 100%, 10.4%)',
-  yellow10: 'hsl(45, 100%, 8.7%)',
-  yellow11: 'hsl(46, 100%, 6.7%)',
-  yellow12: 'hsl(45, 100%, 5.5%)',
-}
-
-export const lightTransparent = 'rgba(255,255,255,0)'
-export const darkTransparent = 'rgba(10,10,10,0)'
-
-export const lightColor = 'hsl(0, 0%, 9.0%)'
-export const lightPalette = [
-  lightTransparent,
-  '#efeff5',
-  '#e6e9f1',
-  '#e5e8f0',
-  '#e0e3ee',
-  '#dde0ec',
-  '#d8dbe8',
-  '#c4cadd',
-  '#a2aecd',
-  '#7689b6',
-  '#627299',
-  '#515e80',
-  lightColor,
-  darkTransparent,
-]
-
-export const darkColor = '#fff'
-export const darkPalette = [
-  darkTransparent,
-  '#12181b',
-  '#192328',
-  '#1d272d',
-  '#263238',
-  '#2b373f',
-  '#32424c',
-  '#41535f',
-  '#475a68',
-  '#516576',
-  '#5c7487',
-  '#7895af',
-  darkColor,
-  lightTransparent,
-]
