--- packages/ui/src/themes/colors-pastel.ts
@@ -1,264 +1,0 @@
-export const blueDark = {
-  blue1: 'hsl(212, 35.0%, 9.2%)',
-  blue2: 'hsl(216, 50.0%, 11.8%)',
-  blue3: 'hsl(214, 59.4%, 15.3%)',
-  blue4: 'hsl(214, 65.8%, 17.9%)',
-  blue5: 'hsl(213, 71.2%, 20.2%)',
-  blue6: 'hsl(212, 77.4%, 23.1%)',
-  blue7: 'hsl(211, 85.1%, 27.4%)',
-  blue8: 'hsl(211, 89.7%, 34.1%)',
-  blue9: 'hsl(206, 100%, 50.0%)',
-  blue10: 'hsl(209, 100%, 60.6%)',
-  blue11: 'hsl(210, 100%, 66.1%)',
-  blue12: 'hsl(206, 98.0%, 95.8%)',
-}
-export const grayDark = {
-  gray1: 'hsl(0, 0%, 8.5%)',
-  gray2: 'hsl(0, 0%, 11.0%)',
-  gray3: 'hsl(0, 0%, 13.6%)',
-  gray4: 'hsl(0, 0%, 15.8%)',
-  gray5: 'hsl(0, 0%, 17.9%)',
-  gray6: 'hsl(0, 0%, 20.5%)',
-  gray7: 'hsl(0, 0%, 24.3%)',
-  gray8: 'hsl(0, 0%, 31.2%)',
-  gray9: 'hsl(0, 0%, 43.9%)',
-  gray10: 'hsl(0, 0%, 49.4%)',
-  gray11: 'hsl(0, 0%, 62.8%)',
-  gray12: 'hsl(0, 0%, 93.0%)',
-}
-export const greenDark = {
-  green1: 'hsl(146, 30.0%, 7.4%)',
-  green2: 'hsl(155, 44.2%, 8.4%)',
-  green3: 'hsl(155, 46.7%, 10.9%)',
-  green4: 'hsl(154, 48.4%, 12.9%)',
-  green5: 'hsl(154, 49.7%, 14.9%)',
-  green6: 'hsl(154, 50.9%, 17.6%)',
-  green7: 'hsl(153, 51.8%, 21.8%)',
-  green8: 'hsl(151, 51.7%, 28.4%)',
-  green9: 'hsl(151, 55.0%, 41.5%)',
-  green10: 'hsl(151, 49.3%, 46.5%)',
-  green11: 'hsl(151, 50.0%, 53.2%)',
-  green12: 'hsl(137, 72.0%, 94.0%)',
-}
-export const orangeDark = {
-  orange1: 'hsl(30, 70.0%, 7.2%)',
-  orange2: 'hsl(28, 100%, 8.4%)',
-  orange3: 'hsl(26, 91.1%, 11.6%)',
-  orange4: 'hsl(25, 88.3%, 14.1%)',
-  orange5: 'hsl(24, 87.6%, 16.6%)',
-  orange6: 'hsl(24, 88.6%, 19.8%)',
-  orange7: 'hsl(24, 92.4%, 24.0%)',
-  orange8: 'hsl(25, 100%, 29.0%)',
-  orange9: 'hsl(24, 94.0%, 50.0%)',
-  orange10: 'hsl(24, 100%, 58.5%)',
-  orange11: 'hsl(24, 100%, 62.2%)',
-  orange12: 'hsl(24, 97.0%, 93.2%)',
-}
-export const pinkDark = {
-  pink1: 'hsl(318, 25.0%, 9.6%)',
-  pink2: 'hsl(319, 32.2%, 11.6%)',
-  pink3: 'hsl(319, 41.0%, 16.0%)',
-  pink4: 'hsl(320, 45.4%, 18.7%)',
-  pink5: 'hsl(320, 49.0%, 21.1%)',
-  pink6: 'hsl(321, 53.6%, 24.4%)',
-  pink7: 'hsl(321, 61.1%, 29.7%)',
-  pink8: 'hsl(322, 74.9%, 37.5%)',
-  pink9: 'hsl(322, 65.0%, 54.5%)',
-  pink10: 'hsl(323, 72.8%, 59.2%)',
-  pink11: 'hsl(325, 90.0%, 66.4%)',
-  pink12: 'hsl(322, 90.0%, 95.8%)',
-}
-export const purpleDark = {
-  purple1: 'hsl(284, 20.0%, 9.6%)',
-  purple2: 'hsl(283, 30.0%, 11.8%)',
-  purple3: 'hsl(281, 37.5%, 16.5%)',
-  purple4: 'hsl(280, 41.2%, 20.0%)',
-  purple5: 'hsl(279, 43.8%, 23.3%)',
-  purple6: 'hsl(277, 46.4%, 27.5%)',
-  purple7: 'hsl(275, 49.3%, 34.6%)',
-  purple8: 'hsl(272, 52.1%, 45.9%)',
-  purple9: 'hsl(272, 51.0%, 54.0%)',
-  purple10: 'hsl(273, 57.3%, 59.1%)',
-  purple11: 'hsl(275, 80.0%, 71.0%)',
-  purple12: 'hsl(279, 75.0%, 95.7%)',
-}
-export const redDark = {
-  red1: 'hsl(353, 23.0%, 9.8%)',
-  red2: 'hsl(357, 34.4%, 12.0%)',
-  red3: 'hsl(356, 43.4%, 16.4%)',
-  red4: 'hsl(356, 47.6%, 19.2%)',
-  red5: 'hsl(356, 51.1%, 21.9%)',
-  red6: 'hsl(356, 55.2%, 25.9%)',
-  red7: 'hsl(357, 60.2%, 31.8%)',
-  red8: 'hsl(358, 65.0%, 40.4%)',
-  red9: 'hsl(358, 75.0%, 59.0%)',
-  red10: 'hsl(358, 85.3%, 64.0%)',
-  red11: 'hsl(358, 100%, 69.5%)',
-  red12: 'hsl(351, 89.0%, 96.0%)',
-}
-export const yellowDark = {
-  yellow1: 'hsl(45, 100%, 5.5%)',
-  yellow2: 'hsl(46, 100%, 6.7%)',
-  yellow3: 'hsl(45, 100%, 8.7%)',
-  yellow4: 'hsl(45, 100%, 10.4%)',
-  yellow5: 'hsl(47, 100%, 12.1%)',
-  yellow6: 'hsl(49, 100%, 14.3%)',
-  yellow7: 'hsl(49, 90.3%, 18.4%)',
-  yellow8: 'hsl(50, 100%, 22.0%)',
-  yellow9: 'hsl(53, 92.0%, 50.0%)',
-  yellow10: 'hsl(54, 100%, 68.0%)',
-  yellow11: 'hsl(48, 100%, 47.0%)',
-  yellow12: 'hsl(53, 100%, 91.0%)',
-}
-
-export const blue = {
-  blue1: 'hsl(206, 98.0%, 95.8%)',
-  blue2: 'hsl(210, 100%, 66.1%)',
-  blue3: 'hsl(209, 100%, 60.6%)',
-  blue4: 'hsl(206, 100%, 50.0%)',
-  blue5: 'hsl(211, 89.7%, 34.1%)',
-  blue6: 'hsl(211, 85.1%, 27.4%)',
-  blue7: 'hsl(212, 77.4%, 23.1%)',
-  blue8: 'hsl(213, 71.2%, 20.2%)',
-  blue9: 'hsl(214, 65.8%, 17.9%)',
-  blue10: 'hsl(214, 59.4%, 15.3%)',
-  blue11: 'hsl(216, 50.0%, 11.8%)',
-  blue12: 'hsl(212, 35.0%, 9.2%)',
-}
-export const gray = {
-  gray1: 'hsl(0, 0%, 93.0%)',
-  gray2: 'hsl(0, 0%, 62.8%)',
-  gray3: 'hsl(0, 0%, 49.4%)',
-  gray4: 'hsl(0, 0%, 43.9%)',
-  gray5: 'hsl(0, 0%, 31.2%)',
-  gray6: 'hsl(0, 0%, 24.3%)',
-  gray7: 'hsl(0, 0%, 20.5%)',
-  gray8: 'hsl(0, 0%, 17.9%)',
-  gray9: 'hsl(0, 0%, 15.8%)',
-  gray10: 'hsl(0, 0%, 13.6%)',
-  gray11: 'hsl(0, 0%, 11.0%)',
-  gray12: 'hsl(0, 0%, 8.5%)',
-}
-export const green = {
-  green1: 'hsl(137, 72.0%, 94.0%)',
-  green2: 'hsl(151, 50.0%, 53.2%)',
-  green3: 'hsl(151, 49.3%, 46.5%)',
-  green4: 'hsl(151, 55.0%, 41.5%)',
-  green5: 'hsl(151, 51.7%, 28.4%)',
-  green6: 'hsl(153, 51.8%, 21.8%)',
-  green7: 'hsl(154, 50.9%, 17.6%)',
-  green8: 'hsl(154, 49.7%, 14.9%)',
-  green9: 'hsl(154, 48.4%, 12.9%)',
-  green10: 'hsl(155, 46.7%, 10.9%)',
-  green11: 'hsl(155, 44.2%, 8.4%)',
-  green12: 'hsl(146, 30.0%, 7.4%)',
-}
-export const orange = {
-  orange1: 'hsl(24, 97.0%, 93.2%)',
-  orange2: 'hsl(24, 100%, 62.2%)',
-  orange3: 'hsl(24, 100%, 58.5%)',
-  orange4: 'hsl(24, 94.0%, 50.0%)',
-  orange5: 'hsl(25, 100%, 29.0%)',
-  orange6: 'hsl(24, 92.4%, 24.0%)',
-  orange7: 'hsl(24, 88.6%, 19.8%)',
-  orange8: 'hsl(24, 87.6%, 16.6%)',
-  orange9: 'hsl(25, 88.3%, 14.1%)',
-  orange10: 'hsl(26, 91.1%, 11.6%)',
-  orange11: 'hsl(28, 100%, 8.4%)',
-  orange12: 'hsl(30, 70.0%, 7.2%)',
-}
-export const pink = {
-  pink1: 'hsl(322, 90.0%, 95.8%)',
-  pink2: 'hsl(325, 90.0%, 66.4%)',
-  pink3: 'hsl(323, 72.8%, 59.2%)',
-  pink4: 'hsl(322, 65.0%, 54.5%)',
-  pink5: 'hsl(322, 74.9%, 37.5%)',
-  pink6: 'hsl(321, 61.1%, 29.7%)',
-  pink7: 'hsl(321, 53.6%, 24.4%)',
-  pink8: 'hsl(320, 49.0%, 21.1%)',
-  pink9: 'hsl(320, 45.4%, 18.7%)',
-  pink10: 'hsl(319, 41.0%, 16.0%)',
-  pink11: 'hsl(319, 32.2%, 11.6%)',
-  pink12: 'hsl(318, 25.0%, 9.6%)',
-}
-export const purple = {
-  purple1: 'hsl(279, 75.0%, 95.7%)',
-  purple2: 'hsl(275, 80.0%, 71.0%)',
-  purple3: 'hsl(273, 57.3%, 59.1%)',
-  purple4: 'hsl(272, 51.0%, 54.0%)',
-  purple5: 'hsl(272, 52.1%, 45.9%)',
-  purple6: 'hsl(275, 49.3%, 34.6%)',
-  purple7: 'hsl(277, 46.4%, 27.5%)',
-  purple8: 'hsl(279, 43.8%, 23.3%)',
-  purple9: 'hsl(280, 41.2%, 20.0%)',
-  purple10: 'hsl(281, 37.5%, 16.5%)',
-  purple11: 'hsl(283, 30.0%, 11.8%)',
-  purple12: 'hsl(284, 20.0%, 9.6%)',
-}
-export const red = {
-  red1: 'hsl(351, 89.0%, 96.0%)',
-  red2: 'hsl(358, 100%, 69.5%)',
-  red3: 'hsl(358, 85.3%, 64.0%)',
-  red4: 'hsl(358, 75.0%, 59.0%)',
-  red5: 'hsl(358, 65.0%, 40.4%)',
-  red6: 'hsl(357, 60.2%, 31.8%)',
-  red7: 'hsl(356, 55.2%, 25.9%)',
-  red8: 'hsl(356, 51.1%, 21.9%)',
-  red9: 'hsl(356, 47.6%, 19.2%)',
-  red10: 'hsl(356, 43.4%, 16.4%)',
-  red11: 'hsl(357, 34.4%, 12.0%)',
-  red12: 'hsl(353, 23.0%, 9.8%)',
-}
-export const yellow = {
-  yellow1: 'hsl(53, 100%, 91.0%)',
-  yellow2: 'hsl(48, 100%, 47.0%)',
-  yellow3: 'hsl(54, 100%, 68.0%)',
-  yellow4: 'hsl(53, 92.0%, 50.0%)',
-  yellow5: 'hsl(50, 100%, 22.0%)',
-  yellow6: 'hsl(49, 90.3%, 18.4%)',
-  yellow7: 'hsl(49, 100%, 14.3%)',
-  yellow8: 'hsl(47, 100%, 12.1%)',
-  yellow9: 'hsl(45, 100%, 10.4%)',
-  yellow10: 'hsl(45, 100%, 8.7%)',
-  yellow11: 'hsl(46, 100%, 6.7%)',
-  yellow12: 'hsl(45, 100%, 5.5%)',
-}
-
-export const lightTransparent = 'rgba(255,255,255,0)'
-export const darkTransparent = 'rgba(10,10,10,0)'
-
-export const lightColor = 'hsl(0, 0%, 9.0%)'
-export const lightPalette = [
-  lightTransparent,
-  '#efeff1',
-  '#e8e9ec',
-  '#e7e8eb',
-  '#e2e3e7',
-  '#dfe0e5',
-  '#dadbe0',
-  '#c8cad1',
-  '#afb3be',
-  '#7b8192',
-  '#6d7281',
-  '#5a5e6b',
-  lightColor,
-  darkTransparent,
-]
-
-export const darkColor = '#fff'
-export const darkPalette = [
-  darkTransparent,
-  '#050505',
-  '#141516',
-  '#18191a',
-  '#222324',
-  '#27282a',
-  '#303234',
-  '#404244',
-  '#47494b',
-  '#525457',
-  '#5f6265',
-  '#a1a6aa',
-  darkColor,
-  lightTransparent,
-]
