+++ packages/ui/src/themes/colors.ts
@@ -1,12 +1,62 @@
-/**
- * change this export to update the colors
- *
- * 1. './colors-default' - the default tamagui theme
- * 2. './colors-neon' - based on the default theme, less saturated
- * 2. './colors-pastel' - based on the default theme, more saturated
- */
+export {
+  blue,
+  blueDark,
+  gray,
+  grayDark,
+  green,
+  greenDark,
+  orange,
+  orangeDark,
+  pink,
+  pinkDark,
+  purple,
+  purpleDark,
+  red,
+  redDark,
+  yellow,
+  yellowDark,
+  gold,
+  goldDark,
+} from '@tamagui/themes'
 
-export * from './colors-default'
+export const lightTransparent = 'rgba(255,255,255,0)'
+export const darkTransparent = 'rgba(10,10,10,0)'
+
+export const lightColor = 'hsl(0, 0%, 9.0%)'
+export const lightPalette = [
+  lightTransparent,
+  '#fff',
+  '#f9f9f9',
+  'hsl(0, 0%, 97.3%)',
+  'hsl(0, 0%, 95.1%)',
+  'hsl(0, 0%, 94.0%)',
+  'hsl(0, 0%, 92.0%)',
+  'hsl(0, 0%, 89.5%)',
+  'hsl(0, 0%, 81.0%)',
+  'hsl(0, 0%, 56.1%)',
+  'hsl(0, 0%, 50.3%)',
+  'hsl(0, 0%, 42.5%)',
+  lightColor,
+  darkTransparent,
+]
+
+export const darkColor = '#fff'
+export const darkPalette = [
+  darkTransparent,
+  '#050505',
+  '#151515',
+  '#191919',
+  '#232323',
+  '#282828',
+  '#323232',
+  '#424242',
+  '#494949',
+  '#545454',
+  '#626262',
+  '#a5a5a5',
+  darkColor,
+  lightTransparent,
+]
 
 export const brandColors = {
   primary: '#C3AB8E',
